feat: allow decimal input across all number fields
Added step="any" to every type="number" input that was missing it
(browsers default to step=1 for type="number", silently rounding or
rejecting decimals otherwise) — product quantity/avgWeightGrams,
recipe ingredient quantity/yieldWeight/portionWeight/salePrice,
inline product create, stock movement quantity.

Also dropped .int() from products.schema.js's quantity — stock is
now numeric(10,2) on the backend (see back-end commit), so weight-
based products (kg, l) can carry fractional stock.

+++ src/features/products/products.schema.js
@@ -3,7 +3,7 @@ import { z } from "zod"
 export const productSchema = z.object({
   name: z.string().min(1, "Name is required"),
   price: z.coerce.number().positive("Price must be greater than 0"),
-  quantity: z.coerce.number().int().nonnegative("Quantity can't be negative"),
+  quantity: z.coerce.number().nonnegative("Quantity can't be negative"),
   unit: z.enum(["g", "kg", "ml", "l", "pcs"]),
   avgWeightGrams: z.preprocess((val) => {
     if(val === "" || val === null){
